test: move service dataset coverage to unit tests
